refactor: 영화 별점 기능 및 코멘트 분리 및 폴더 구조 변경
